criado aplicação para alugar carro

+++ Locadora/src/main/java/com/locadora/dto/veiculo/DtoVeiculo.java
@@ -14,7 +14,7 @@ public class DtoVeiculo {
 	public String modelo;
 	public String cor;
 	public String placa;
-	public BigDecimal valorDiaria;
+	public double valorDiaria;
 	public boolean disponivel;
 	
 }
